% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_new_chapter2.R
\name{set_new_chapter2}
\alias{set_new_chapter2}
\title{Insert equals from courser position to up to 80 characters}
\usage{
set_new_chapter2()
}
\value{
equals inside RStudio
}
\description{
Insert equals from courser position to up to 80 characters
}
